4.21*: Blog List Expansion, step 9

+++ part4/blog/controllers/blogs.js
@@ -37,8 +37,19 @@ blogsRouter.post('/', tokenExtractor, async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  if(user.id.toString() === blog.user.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }else{
+    response.status(400).end()
+  }
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
